+++ src/index.ts
@@ -1,6 +1,6 @@
 export * from "./createPat.js";
 export * from "./createUserNpmrc.js";
-export * from "./makeParsedProjectNpmrc.js";
-export * from "./parseProjectNpmrc.js";
+export * from "./projectNpmrcMake.js";
+export * from "./projectNpmrcParse.js";
 export * from "./types.js";
 export * from "./writeNpmrc.js";
