+++ src/projectNpmrcParse.ts
